Remove old JUnit multiversion spec and test rewriter
